[JBPM-9255] Fix TaskEvent.processType for validation at some dbs (#1710)

+++ jbpm-db-scripts/src/main/resources/db/ddl-scripts/sqlserver/sqlserver-jbpm-schema.sql
@@ -414,7 +414,7 @@
         OPTLOCK int,
         workItemId numeric(19,0),
         correlationKey varchar(255),
-        processType numeric(1,0),
+        processType int,
         primary key (id)
     );
